+++ tfe/resource_tfe_organization_membership.go
@@ -4,8 +4,10 @@
 package tfe
 
 import (
+	"context"
 	"fmt"
 	"log"
+	"strings"
 
 	tfe "github.com/hashicorp/go-tfe"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -17,7 +19,7 @@ func resourceTFEOrganizationMembership() *schema.Resource {
 		Read:   resourceTFEOrganizationMembershipRead,
 		Delete: resourceTFEOrganizationMembershipDelete,
 		Importer: &schema.ResourceImporter{
-			StateContext: schema.ImportStatePassthroughContext,
+			StateContext: resourceTFEOrganizationMembershipImporter,
 		},
 
 		Schema: map[string]*schema.Schema{
@@ -114,4 +116,26 @@ func resourceTFEOrganizationMembershipDelete(d *schema.ResourceData, meta interf
 	}
 
 	return nil
+}
+
+func resourceTFEOrganizationMembershipImporter(ctx context.Context, d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+	config := meta.(ConfiguredClient)
+
+	// Import formats:
+	//  - <ORGANIZATION MEMBERSHIP ID>
+	//  - <organization name>/<user email>
+	s := strings.SplitN(d.Id(), "/", 2)
+	if len(s) == 2 {
+		org := s[0]
+		email := s[1]
+		orgMembership, err := fetchOrganizationMemberByNameOrEmail(ctx, config.Client, org, "", email)
+		if err != nil {
+			return nil, fmt.Errorf(
+				"error retrieving user with email %s from organization %s: %w", email, org, err)
+		}
+
+		d.SetId(orgMembership.ID)
+	}
+
+	return []*schema.ResourceData{d}, nil
 }